Change delta time type into examples.

+++ example/assimp/src/Main.cpp
@@ -280,7 +280,7 @@ main(int argc, char** argv)
 		}
 	});
 
-	auto enterFrame = canvas->enterFrame()->connect([&](Canvas::Ptr canvas, uint time, uint deltaTime)
+	auto enterFrame = canvas->enterFrame()->connect([&](Canvas::Ptr canvas, uint time, float deltaTime)
 	{
 		sceneManager->nextFrame();
 	});
